+++ editor/nvim/lua/plugins/yoga-js.lua
@@ -1,0 +1,25 @@
+-- Yoga Files: JS/TS defaults for LazyVim
+
+return {
+  {
+    "williamboman/mason.nvim",
+    opts = function(_, opts)
+      opts.ensure_installed = opts.ensure_installed or {}
+      vim.list_extend(opts.ensure_installed, {
+        "typescript-language-server",
+        "biome",
+      })
+    end,
+  },
+  {
+    "stevearc/conform.nvim",
+    opts = function(_, opts)
+      opts.formatters_by_ft = opts.formatters_by_ft or {}
+      opts.formatters_by_ft.javascript = { "biome" }
+      opts.formatters_by_ft.javascriptreact = { "biome" }
+      opts.formatters_by_ft.typescript = { "biome" }
+      opts.formatters_by_ft.typescriptreact = { "biome" }
+      opts.formatters_by_ft.json = { "biome" }
+    end,
+  },
+}